docs(README): clarify Sakata Index scope and improve Node.js explanation
This commit updates the README.md to provide clearer information on the Sakata Index, specifying that it is designed for the Nikkei 225. It also refines the explanation regarding the choice of Node.js 16 and the optimization made for dot product calculations. These changes aim to enhance understanding for readers about the specific application of the Sakata Index and the technical decisions behind the project's development.

- Specified that the Sakata Index uses a neural network designed for the Nikkei 225, making its application clearer.
- Improved clarity on why Node.js 16 was chosen, emphasizing the speed improvement in dot product calculations by substituting mathjs's 'dot' function with 'reduce', which results in faster learning speeds compared to using brain.js.